requirements.txt
#dependencies
requests
pandas
numpy

# Loads variables from .env
python-dotenv